Refactor to new package structure.
