+++ src/main/java/me/borawski/hcf/command/commands/FriendsCommand.java
@@ -1,11 +1,11 @@
 package me.borawski.hcf.command.commands;
 
 import me.borawski.hcf.command.CustomBaseCommand;
-import me.borawski.hcf.command.commands.friends.FriendAcceptCommand;
-import me.borawski.hcf.command.commands.friends.FriendAddCommand;
-import me.borawski.hcf.command.commands.friends.FriendDeclineCommand;
-import me.borawski.hcf.command.commands.friends.FriendListCommand;
-import me.borawski.hcf.command.commands.friends.FriendRemoveCommand;
+import me.borawski.hcf.command.commands.sub.FriendAcceptCommand;
+import me.borawski.hcf.command.commands.sub.FriendAddCommand;
+import me.borawski.hcf.command.commands.sub.FriendDeclineCommand;
+import me.borawski.hcf.command.commands.sub.FriendListCommand;
+import me.borawski.hcf.command.commands.sub.FriendRemoveCommand;
 import me.borawski.hcf.session.Rank;
 
 /**
