+++ src/main/java/me/borawski/hcf/command/commands/sub/FriendDeclineCommand.java
@@ -1,4 +1,4 @@
-package me.borawski.hcf.command.commands.friends;
+package me.borawski.hcf.command.commands.sub;
 
 import org.bukkit.ChatColor;
 import org.bukkit.command.CommandSender;
@@ -13,10 +13,10 @@ import me.borawski.hcf.util.ChatUtils;
 import me.borawski.hcf.util.FriendUtils;
 import me.borawski.hcf.util.PlayerUtils;
 
-public class FriendAcceptCommand extends CustomCommand {
+public class FriendDeclineCommand extends CustomCommand {
 
-    public FriendAcceptCommand() {
-        super("accept", "Accept a friend request.", Rank.GUEST, "confirm");
+    public FriendDeclineCommand() {
+        super("decline", "Decline a friend request.", Rank.GUEST, "deny");
     }
 
     @Override
@@ -38,19 +38,13 @@ public class FriendAcceptCommand extends CustomCommand {
             }
 
             if (!FriendUtils.hasRequest(session, target.getUniqueId())) {
-                // Sender has not already sent a friend request to this
-                // player. //
-                sender.sendMessage(Core.getInstance().getPrefix() + ChatColor.GRAY + "You don't have a request from " + ChatUtils.getNameWithRankColor(target.getUniqueId(), false));
+                sender.sendMessage(Core.getInstance().getPrefix() + ChatColor.GRAY + "You don't have a request from " + ChatUtils.getNameWithRankColor(target.getUniqueId(), false) + ChatColor.GRAY + "!");
                 return;
             }
 
-            sender.sendMessage(Core.getInstance().getPrefix() + ChatColor.GRAY + "You accepted a request from " + ChatUtils.getNameWithRankColor(target.getUniqueId(), false));
-            sender.sendMessage(Core.getInstance().getPrefix() + ChatColor.GRAY + "You are now friends with " + ChatUtils.getNameWithRankColor(target.getUniqueId(), false) + ChatColor.GRAY + "!");
-
-            FriendUtils.acceptFriendRequest(session, target.getUniqueId(), true);
-            FriendUtils.acceptFriendRequest(target, session.getUniqueId(), false);
-        } else {
-            sender.sendMessage(Core.getLangHandler().getString("usage-header") + "§7/friend " + label + " [player]");
+            FriendUtils.denyFriendRequest(session, target.getUniqueId(), true);
+            FriendUtils.denyFriendRequest(target, session.getUniqueId(), false);
+            sender.sendMessage(Core.getInstance().getPrefix() + ChatColor.GRAY + "You denied " + ChatUtils.getNameWithRankColor(target.getUniqueId(), false) + ChatColor.GRAY + "'s friend invite!");
         }
     }
