+++ src/main/java/me/borawski/hcf/command/commands/sub/RankListCommand.java
@@ -1,4 +1,4 @@
-package me.borawski.hcf.command.commands.rank;
+package me.borawski.hcf.command.commands.sub;
 
 import org.bukkit.command.CommandSender;
 
@@ -6,15 +6,19 @@ import me.borawski.hcf.api.RankAPI;
 import me.borawski.hcf.command.CustomCommand;
 import me.borawski.hcf.session.Rank;
 
-public class RankCheckCommand extends CustomCommand {
+/**
+ * @author Ryan Radomski
+ * 
+ */
+public class RankListCommand extends CustomCommand {
 
-    public RankCheckCommand() {
-        super("check", "Check your rank.", Rank.GUEST);
+    public RankListCommand() {
+        super("list", "list all the ranks", Rank.GUEST);
     }
 
     @Override
     public void run(CommandSender sender, String label, String[] args) {
-        RankAPI.checkRank(sender, label);
+        RankAPI.listRanks(sender);
     }
 
 }
