+++ src/main/java/me/borawski/hcf/command/commands/sub/RankSetCommand.java
@@ -1,4 +1,4 @@
-package me.borawski.hcf.command.commands.rank;
+package me.borawski.hcf.command.commands.sub;
 
 import org.bukkit.command.CommandSender;
